requirements.txt
requests==2.28.1
grafana-api==1.0.3

PyYAML~=6.0